requirements.txt
numpy
pandas
fastapi
uvicorn[standard]
streamlit
openai
python-dotenv
scikit-learn
